+++ src/pam_analyzer/infrastructure/__init__.py
@@ -2,7 +2,6 @@ from .analysis_discovery import discover_analysis_result
 from .audio_extractor import SoundfileAudioExtractor
 from .audio_import import AudioImporter
 from .audio_inventory_discovery import discover_audio_inventory
-from .birdnet_analyzer import BirdnetAnalyzer
 from .birdnet_runner import BirdnetRunner
 from .csv_detection_repo import CsvDetectionRepository
 from .psutil_sdcard_scanner import PsutilSdCardScanner
@@ -11,7 +10,6 @@ from .toml_project_repo import TomlProjectRepository
 
 __all__ = [
     "AudioImporter",
-    "BirdnetAnalyzer",
     "BirdnetRunner",
     "CsvDetectionRepository",
     "PsutilSdCardScanner",